FEAT: Connect to the server and test POST
Implemented the connection between the app and the server. Connected via
Ktor and checked the POST is working well by implementing simple login
screen, writing necessary informations in each blanks, and then touching
the button. The result is successful, it shows the Toast Message that it
was successful and the MySQL Database is also showing that the new data
has posted.